Consolidating test data folders
