+++ tests/data/model_assembly/GrFN/PETPT--GrFN.json
