--- tests/data/model_assembly/GrFN_execution/example_input.json
@@ -1,24 +1,0 @@
-{
-    "PID::@global.PID_realize::speed::0": [
-        10,
-        10,
-        10,
-        20,
-        20,
-        20,
-        30,
-        30,
-        30
-    ],
-    "PID::@global.PID_init::pid.Kp::0": [
-        0.15,
-        0.2,
-        0.25,
-        0.15,
-        0.2,
-        0.25,
-        0.15,
-        0.2,
-        0.25
-    ]
-}
